Moved Shared ValueObjects into Shared
