+++ FamilyPets.Backend/src/FamilyPets.Domain/SpeciesAgregate/BreedId.cs
@@ -1,5 +1,6 @@
 ﻿using FamilyForPets.Domain.VolunteerAgregate;
 using FamilyForPets.Domain.VolunteerAgregate.PetValueObjects;
+using FamilyForPets.Shared.ValueObjects.IDs;
 
 namespace FamilyForPets.Domain.SpeciesAgregate
 {
