+++ FamilyPets.Backend/src/FamilyPets.Domain/VolunteerAgregate/PetValueObjects/PetId.cs
@@ -1,5 +1,4 @@
-﻿using CSharpFunctionalExtensions;
-using FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects;
+﻿using FamilyForPets.Shared.ValueObjects.IDs;
 
 namespace FamilyForPets.Domain.VolunteerAgregate.PetValueObjects
 {
