+++ FamilyPets.Backend/src/FamilyPets.Domain/VolunteerAgregate/PetValueObjects/PetVaccinesList.cs
@@ -1,5 +1,4 @@
 ﻿using CSharpFunctionalExtensions;
-using FamilyForPets.Domain.SharedValueObjects;
 using FamilyForPets.Shared;
 
 namespace FamilyForPets.Domain.VolunteerAgregate.PetValueObjects
