+++ FamilyPets.Backend/src/FamilyPets.Infrastructure/Configurations/PetConfiguration.cs
@@ -1,10 +1,10 @@
-﻿using FamilyForPets.Domain.SharedValueObjects;
-using FamilyForPets.Domain.SpeciesAgregate;
+﻿using FamilyForPets.Domain.SpeciesAgregate;
 using FamilyForPets.Domain.VolunteerAgregate;
 using FamilyForPets.Domain.VolunteerAgregate.PetValueObjects;
 using FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects;
 using FamilyForPets.Infrastructure.Configurations.Converters;
 using FamilyForPets.Shared;
+using FamilyForPets.Shared.ValueObjects;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
