+++ FamilyPets.Backend/src/FamilyPets.Infrastructure/Repositories/VolunteersRepository.cs
@@ -1,8 +1,8 @@
 ﻿using CSharpFunctionalExtensions;
-using FamilyForPets.Domain.SharedValueObjects;
 using FamilyForPets.Domain.VolunteerAgregate;
 using FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects;
 using FamilyForPets.Shared;
+using FamilyForPets.Shared.ValueObjects;
 using FamilyForPets.UseCases.VolunteerAgregate;
 using Microsoft.EntityFrameworkCore;
