+++ FamilyPets.Backend/src/FamilyPets.Shared/ValueObjects/Adress.cs
@@ -1,7 +1,7 @@
 ﻿using CSharpFunctionalExtensions;
 using FamilyForPets.Shared;
 
-namespace FamilyForPets.Domain.SharedValueObjects
+namespace FamilyForPets.Shared.ValueObjects
 {
     public class Adress : ValueObject
     {
