+++ FamilyPets.Backend/src/FamilyPets.Shared/ValueObjects/Length.cs
@@ -1,7 +1,7 @@
 ﻿using CSharpFunctionalExtensions;
 using FamilyForPets.Shared;
 
-namespace FamilyForPets.Domain.SharedValueObjects
+namespace FamilyForPets.Shared.ValueObjects
 {
     public class Length : ValueObject
     {
