+++ FamilyPets.Backend/src/FamilyPets.Shared/ValueObjects/PhoneNumber.cs
@@ -2,7 +2,7 @@
 using CSharpFunctionalExtensions;
 using FamilyForPets.Shared;
 
-namespace FamilyForPets.Domain.SharedValueObjects
+namespace FamilyForPets.Shared.ValueObjects
 {
     public class PhoneNumber : ValueObject
     {
