+++ FamilyPets.Backend/src/FamilyPets.UseCases/VolunteerAgregate/IVolunteerRepository.cs
@@ -1,8 +1,8 @@
 ﻿using CSharpFunctionalExtensions;
-using FamilyForPets.Domain.SharedValueObjects;
 using FamilyForPets.Domain.VolunteerAgregate;
 using FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects;
 using FamilyForPets.Shared;
+using FamilyForPets.Shared.ValueObjects;
 
 namespace FamilyForPets.UseCases.VolunteerAgregate
 {
