+++ FamilyPets.Backend/src/FamilyPets.UseCases/VolunteerAgregate/UpdateVolunteer/UpdateVolunteerCommandValidator.cs
@@ -1,6 +1,6 @@
-﻿using FamilyForPets.Domain.SharedValueObjects;
-using FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects;
+﻿using FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects;
 using FamilyForPets.Shared;
+using FamilyForPets.Shared.ValueObjects;
 using FamilyForPets.UseCases.Validation;
 using FluentValidation;
