+++ FamilyPets.Backend/src/FamilyPets.UseCases/VolunteerAgregate/UpdateVolunteer/UpdateVolunteerDetailsForPayment/UpdateVolunteerDetailsForPaymentHandler.cs
@@ -1,11 +1,9 @@
 ﻿using CSharpFunctionalExtensions;
-using FamilyForPets.Domain.SharedValueObjects;
 using FamilyForPets.Domain.VolunteerAgregate;
 using FamilyForPets.Domain.VolunteerAgregate.VolunteerValueObjects;
 using FamilyForPets.Shared;
 using FamilyForPets.UseCases.Abstractions;
 using FamilyForPets.UseCases.Extentions.ValidationExtentions;
-using FamilyForPets.UseCases.VolunteerAgregate.UpdateVolunteer.UpdateVolunteerSocialNetworks;
 using FluentValidation;
 using FluentValidation.Results;
 using Microsoft.Extensions.Logging;
