+++ FamilyPets.Backend/src/FamilyPets.UseCases/VolunteerAgregate/UpdateVolunteer/UpdateVolunteerMainInfo/UpdateVolunteerMainInfoCommandValidator.cs
@@ -1,4 +1,4 @@
-﻿using FamilyForPets.Domain.SharedValueObjects;
+﻿using FamilyForPets.Shared.ValueObjects;
 using FamilyForPets.UseCases.Validation;
 using FluentValidation;
